修改bug

+++ app/src/main/java/com/kssoft/lake/data/model/commit/XcLakeR.java
@@ -56,7 +56,7 @@ public class XcLakeR extends QualityStandard implements JSON, Serializable {
     private String gaqp;
     private String gaqpam;
     private Double ph;
-    private Integer redox;
+    private Double redox;
     private String smell;
 
     private Double wndv;
@@ -161,7 +161,7 @@ public class XcLakeR extends QualityStandard implements JSON, Serializable {
         return this.ph;
     }
 
-    public Integer getRedox() {
+    public Double getRedox() {
         return this.redox;
     }
 
@@ -291,7 +291,7 @@ public class XcLakeR extends QualityStandard implements JSON, Serializable {
         this.ph = var1;
     }
 
-    public void setRedox(Integer var1) {
+    public void setRedox(Double var1) {
         this.redox = var1;
     }
